split filters into multiple files
