+++ graphqlapi/local/common_filters/ast.go
@@ -1,0 +1,46 @@
+package common_filters
+
+import (
+	"github.com/creativesoftwarefdn/weaviate/database/schema"
+)
+
+type Operator int
+
+const (
+	OperatorEqual            Operator = 1
+	OperatorNotEqual         Operator = 2
+	OperatorGreaterThan      Operator = 3
+	OperatorGreaterThanEqual Operator = 4
+	OperatorLessThan         Operator = 5
+	OperatorLessThanEqual    Operator = 6
+	OperatorAnd              Operator = 7
+	OperatorOr               Operator = 8
+	OperatorNot              Operator = 9
+)
+
+type LocalFilter struct {
+	Root *Clause
+}
+
+// Represents the path in a filter.
+// Either RelationProperty or PrimitiveProperty must be empty (e.g. "").
+type Path struct {
+	Class    schema.ClassName
+	Property schema.PropertyName
+
+	// If nil, than this is the property we're interested in.
+	// If a pointer to another Path, the constraint applies to that one.
+	Child *Path
+}
+
+type Value struct {
+	Value interface{}
+	Type  schema.DataType
+}
+
+type Clause struct {
+	Operator Operator
+	On       *Path
+	Value    *Value
+	Operands []Clause
+}
